did some small renaming of iCloudFull to iCloudFullLevel to match the naming convention of iCloudTopLevel - helps clarify the difference with the Main scripts. iCloudFullMain, the initial comment at the start was changed from its original copy paste of the TopMain comment explaining the script's function.
